Start Project
1. Made constructor function to make sliders
2. sliderTest.html is used to test constructor functionality
temporarily.
3. constructor is passed id of div that I want to become the slider -
this div is given the class 'slider', then a content and two (left and
right) nav buttons are given classes and added to the DOM inside this
slider div.
4. 'testJSON' is an object in the js file created to avoid and AJAX call
for testing purposes.
5. this object should have a property that matches the id of the
original div.
6. i have a queryData function to check if that property is in fact
there, if so a 'useData' function is called that will be added to in the
future, if the property is not in the main JSON object an alert appears
telling the user to update the json object.